[DoNotCarryForward] Reland "Move EventAutoRepeatHandler out of /evdev umbrella"
This relands commit https://crrev.com/881961 reverted by
https://crrev.com/c/882721 due to
https://ci.chromium.org/buildbot/chromium.chromiumos/linux-chromeos-dbg/3890

It fixes a missing export that broken component builds of ChromeOS/Ozone.

TBR=spang@chromium.org

BUG=578890

Reviewed-on: https://chromium-review.googlesource.com/883581
Commit-Queue: Antonio Gomes <tonikitoo@igalia.com>
Cr-Commit-Position: refs/heads/master@{#531519}
