+++ ui/events/ozone/BUILD.gn
@@ -33,6 +33,8 @@ if (use_ozone) {
       "device/udev/device_manager_udev.cc",
       "device/udev/device_manager_udev.h",
       "events_ozone_export.h",
+      "keyboard/event_auto_repeat_handler.cc",
+      "keyboard/event_auto_repeat_handler.h",
     ]
 
     deps = [
@@ -70,8 +72,6 @@ if (use_ozone) {
     sources = [
       "evdev/device_event_dispatcher_evdev.cc",
       "evdev/device_event_dispatcher_evdev.h",
-      "evdev/event_auto_repeat_handler.cc",
-      "evdev/event_auto_repeat_handler.h",
       "evdev/event_converter_evdev.cc",
       "evdev/event_converter_evdev.h",
       "evdev/event_converter_evdev_impl.cc",
